Refactor the defaults into helper files to lay groundwork for version controlled user-side storage
